# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Campus Faso - Gestion du LMD",
        "description": "Documentation générée par Swagger",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/ufrs": {
            "get": {
                "tags": [
                    "UFR"
                ],
                "summary": "Récupérer la liste de toutes les UFR",
                "operationId": "388cfebe285d5023451b2a218c2e14f5",
                "responses": {
                    "200": {
                        "description": "Réponse de succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "universite_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "libelle": {
                                                "type": "string",
                                                "example": "UFR Sciences et Techniques"
                                            },
                                            "abreviation": {
                                                "type": "string",
                                                "example": "UFR/ST"
                                            },
                                            "responsable_id": {
                                                "type": "string",
                                                "example": "12345"
                                            },
                                            "contact": {
                                                "type": "string",
                                                "example": "+226 25 30 70 65"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "ufrst@uo.bf"
                                            },
                                            "statut": {
                                                "type": "string",
                                                "example": "active"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/ufrs/{id}": {
            "get": {
                "tags": [
                    "UFR"
                ],
                "summary": "Récupérer une UFR par son ID",
                "operationId": "714a3611a6fec515a59688f7a008d844",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'UFR",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réponse de succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "universite_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "libelle": {
                                                "type": "string",
                                                "example": "UFR Sciences et Techniques"
                                            },
                                            "abreviation": {
                                                "type": "string",
                                                "example": "UFR/ST"
                                            },
                                            "responsable_id": {
                                                "type": "string",
                                                "example": "12345"
                                            },
                                            "contact": {
                                                "type": "string",
                                                "example": "+226 25 30 70 65"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "ufrst@uo.bf"
                                            },
                                            "statut": {
                                                "type": "string",
                                                "example": "active"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "UFR non trouvée"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/ufrs/{ufr}/departements": {
            "get": {
                "tags": [
                    "UFR"
                ],
                "summary": "Récupérer la liste des départements appartenant à une UFR",
                "operationId": "4a27bddad3a633ea2473ed8a4bb4c25d",
                "parameters": [
                    {
                        "name": "ufr",
                        "in": "path",
                        "description": "ID de l'UFR",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réponse de succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "ufr_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "libelle": {
                                                "type": "string",
                                                "example": "Département de Mathématiques"
                                            },
                                            "abreviation": {
                                                "type": "string",
                                                "example": "MATH"
                                            },
                                            "responsable_id": {
                                                "type": "string",
                                                "example": "12346"
                                            },
                                            "contact": {
                                                "type": "string",
                                                "example": "+226 25 30 70 66"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "math@uo.bf"
                                            },
                                            "statut": {
                                                "type": "string",
                                                "example": "active"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "UFR non trouvée"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/ufrs/{ufr}/filieres": {
            "get": {
                "tags": [
                    "UFR"
                ],
                "summary": "Récupérer la liste des filières appartenant à une UFR",
                "operationId": "e242b0c2489b6b229d9b8276fefa9938",
                "parameters": [
                    {
                        "name": "ufr",
                        "in": "path",
                        "description": "ID de l'UFR",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réponse de succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "departement_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "libelle": {
                                                "type": "string",
                                                "example": "Informatique"
                                            },
                                            "abreviation": {
                                                "type": "string",
                                                "example": "INFO"
                                            },
                                            "responsable_id": {
                                                "type": "string",
                                                "example": "12347"
                                            },
                                            "contact": {
                                                "type": "string",
                                                "example": "+226 25 30 70 67"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "info@uo.bf"
                                            },
                                            "statut": {
                                                "type": "string",
                                                "example": "active"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "UFR non trouvée"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/ufrs/{ufr}/niveaux": {
            "get": {
                "tags": [
                    "UFR"
                ],
                "summary": "Récupérer la liste des niveaux appartenant à une UFR",
                "operationId": "d4904c6b55c274367ef024a22c406d0d",
                "parameters": [
                    {
                        "name": "ufr",
                        "in": "path",
                        "description": "ID de l'UFR",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réponse de succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "filiere_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "libelle": {
                                                "type": "string",
                                                "example": "Licence 1"
                                            },
                                            "abreviation": {
                                                "type": "string",
                                                "example": "L1"
                                            },
                                            "accessible": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "statut": {
                                                "type": "string",
                                                "example": "active"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "UFR non trouvée"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/ufrs/{ufr}/niveaux-accessibles": {
            "get": {
                "tags": [
                    "UFR"
                ],
                "summary": "Récupérer la liste des niveaux accessibles appartenant à une UFR",
                "operationId": "411620ef5f8ce5789ff1549dd38df404",
                "parameters": [
                    {
                        "name": "ufr",
                        "in": "path",
                        "description": "ID de l'UFR",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réponse de succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "filiere_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "libelle": {
                                                "type": "string",
                                                "example": "Licence 1"
                                            },
                                            "abreviation": {
                                                "type": "string",
                                                "example": "L1"
                                            },
                                            "accessible": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "statut": {
                                                "type": "string",
                                                "example": "active"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "UFR non trouvée"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
...
    },
...
}